// src/app/general/interfaces/calender.ts
export interface ICalender {
  id: string;
  serviceId: string;
  calenderName: string;
  sunday: boolean;
  monday: boolean;
  tuesday: boolean;
  wednesday: boolean;
  thursday: boolean;
  friday: boolean;
  saturday: boolean;
  startDate: string;
  endDate: string;
  createdAt: string;
  updatedAt: string;
}