// src/app/general/models/calender/calender-dto.ts
/* tslint:disable: variable-name */
export class ReadCalenderDto {
  id: string;
  service_id: string;
  calender_name: string;
  sunday: boolean;
  monday: boolean;
  tuesday: boolean;
  wednesday: boolean;
  thursday: boolean;
  friday: boolean;
  saturday: boolean;
  start_date: string;
  end_date: string;
  created_at: string;
  updated_at: string;
}